+++ src/rollup/gadgets/deposit_block/mod.rs
@@ -2,179 +2,225 @@ use plonky2::{
     field::extension::Extendable,
     hash::hash_types::{HashOut, HashOutTarget, RichField},
     iop::witness::Witness,
-    plonk::{circuit_builder::CircuitBuilder, config::AlgebraicHasher},
+    plonk::{
+        circuit_builder::CircuitBuilder,
+        config::{AlgebraicHasher, Hasher},
+    },
 };
 
 use crate::{
-    sparse_merkle_tree::{
-        gadgets::process::{
-            process_smt::{SmtProcessProof, SparseMerkleProcessProofTarget},
-            utils::{
-                get_process_merkle_proof_role, verify_layered_smt_target_connection,
-                ProcessMerkleProofRoleTarget,
-            },
+    merkle_tree::{gadgets::get_merkle_root_target, tree::KeyLike},
+    transaction::{
+        asset::ContributedAsset,
+        gadgets::{
+            asset_mess::ContributedAssetTarget,
+            purge::{PurgeOutputProcessProof, PurgeOutputProcessProofTarget},
         },
-        layered_tree::verify_layered_smt_connection,
-        proof::ProcessMerkleProofRole,
+        tree::tx_diff::TxDiffTree,
     },
-    utils::hash::WrappedHashOut,
+    utils::gadgets::{
+        hash::poseidon_two_to_one,
+        logic::{conditionally_select, enforce_equal_if_enabled},
+    },
 };
+
+#[allow(clippy::complexity)]
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct DepositBlockProduction<F: RichField, H: Hasher<F>, K: KeyLike> {
+    pub deposit_process_proof: Vec<PurgeOutputProcessProof<F, H, K>>,
+    pub log_n_recipients: usize,
+    pub log_n_kinds: usize,
+}
+
+impl<F: RichField, H: Hasher<F>, K: KeyLike> DepositBlockProduction<F, H, K> {
+    pub fn calculate(&self) -> anyhow::Result<H::Hash> {
+        let deposit_tree = TxDiffTree::<_, H>::new(self.log_n_recipients, self.log_n_kinds);
+
+        let mut prev_deposit_root = deposit_tree.get_root().unwrap();
+        for process_proof in self.deposit_process_proof.iter() {
+            let (old_deposit_root, new_deposit_root) = process_proof.calculate();
+
+            assert_eq!(old_deposit_root, prev_deposit_root);
+
+            prev_deposit_root = new_deposit_root;
+        }
+        dbg!(prev_deposit_root);
+
+        Ok(prev_deposit_root)
+    }
+}
 
 #[derive(Clone, Debug)]
 pub struct DepositBlockProductionTarget {
-    pub deposit_process_proofs: Vec<(
-        SparseMerkleProcessProofTarget,
-        SparseMerkleProcessProofTarget,
-        SparseMerkleProcessProofTarget,
-    )>, // input
+    pub deposit_process_proof: Vec<PurgeOutputProcessProofTarget>, // input
 
     pub interior_deposit_digest: HashOutTarget, // output
 
     pub log_n_recipients: usize, // constant
-    pub log_n_contracts: usize,  // constant
-    pub log_n_variables: usize,  // constant
+    pub log_n_kinds: usize,      // constant
 }
 
 impl DepositBlockProductionTarget {
     pub fn add_virtual_to<F: RichField + Extendable<D>, H: AlgebraicHasher<F>, const D: usize>(
         builder: &mut CircuitBuilder<F, D>,
         log_n_recipients: usize,
-        log_n_contracts: usize,
-        log_n_variables: usize,
-        n_deposits: usize,
+        log_n_kinds: usize,
     ) -> Self {
-        let mut deposit_process_proofs = vec![];
-        for _ in 0..n_deposits {
-            let targets = (
-                SparseMerkleProcessProofTarget::add_virtual_to::<F, H, D>(
+        let n_levels = log_n_recipients + log_n_kinds;
+        let n_deposits = 1 << n_levels;
+        let deposit_process_proof = (0..n_deposits)
+            .map(|_| {
+                PurgeOutputProcessProofTarget::add_virtual_to::<_, H, D>(
                     builder,
                     log_n_recipients,
-                ),
-                SparseMerkleProcessProofTarget::add_virtual_to::<F, H, D>(builder, log_n_contracts),
-                SparseMerkleProcessProofTarget::add_virtual_to::<F, H, D>(builder, log_n_variables),
+                    log_n_kinds,
+                )
+            })
+            .collect::<Vec<_>>();
+
+        let default_asset_target = ContributedAssetTarget::constant_default(builder);
+        let default_leaf_hash = builder.hash_n_to_hash_no_pad::<H>(default_asset_target.encode());
+
+        let default_root_hash = {
+            let mut default_root_hash = default_leaf_hash;
+            for _ in 0..n_levels {
+                default_root_hash =
+                    poseidon_two_to_one::<_, H, D>(builder, default_root_hash, default_root_hash);
+            }
+
+            default_root_hash
+        };
+
+        let mut interior_deposit_digest = default_root_hash;
+        for PurgeOutputProcessProofTarget {
+            siblings: siblings_t,
+            index: index_t,
+            new_leaf_data: new_leaf_data_t,
+            enabled: enabled_t,
+        } in deposit_process_proof.iter()
+        {
+            let proof1_old_leaf_t = default_leaf_hash;
+            let proof1_new_leaf_t = builder.hash_n_to_hash_no_pad::<H>(new_leaf_data_t.encode());
+
+            let proof1_old_root_t =
+                get_merkle_root_target::<F, H, D>(builder, index_t, proof1_old_leaf_t, siblings_t);
+            let proof1_new_root_t =
+                get_merkle_root_target::<F, H, D>(builder, index_t, proof1_new_leaf_t, siblings_t);
+            enforce_equal_if_enabled(
+                builder,
+                interior_deposit_digest,
+                proof1_old_root_t,
+                *enabled_t,
+            );
+            interior_deposit_digest = conditionally_select(
+                builder,
+                proof1_new_root_t,
+                interior_deposit_digest,
+                *enabled_t,
             );
 
-            deposit_process_proofs.push(targets);
+            // deposit する asset の amount が 2^56 未満の値であること
+            builder.range_check(new_leaf_data_t.amount, 56);
         }
 
-        let interior_deposit_digest =
-            calc_deposit_digest::<F, H, D>(builder, &deposit_process_proofs);
-
         Self {
-            deposit_process_proofs,
+            deposit_process_proof,
             interior_deposit_digest,
             log_n_recipients,
-            log_n_contracts,
-            log_n_variables,
+            log_n_kinds,
         }
     }
 
     /// Returns `interior_deposit_digest`
-    pub fn set_witness<F: RichField + Extendable<D>, const D: usize>(
+    pub fn set_witness<F: RichField + Extendable<D>, H: AlgebraicHasher<F>, const D: usize>(
         &self,
         pw: &mut impl Witness<F>,
-        deposit_process_proofs: &[(SmtProcessProof<F>, SmtProcessProof<F>, SmtProcessProof<F>)],
-    ) -> HashOut<F> {
-        let mut prev_interior_deposit_digest = WrappedHashOut::default();
-        assert!(deposit_process_proofs.len() <= self.deposit_process_proofs.len());
-        for (proof_t, proof) in self
-            .deposit_process_proofs
+        value: &DepositBlockProduction<F, H, Vec<bool>>,
+    ) -> anyhow::Result<H::Hash> {
+        for (target, value) in self
+            .deposit_process_proof
             .iter()
-            .zip(deposit_process_proofs.iter())
+            .zip(value.deposit_process_proof.iter())
         {
-            assert_eq!(proof.0.old_root, prev_interior_deposit_digest);
-            verify_layered_smt_connection(
-                proof.0.fnc,
-                proof.0.old_value,
-                proof.0.new_value,
-                proof.1.old_root,
-                proof.1.new_root,
-            )
-            .unwrap();
-            verify_layered_smt_connection(
-                proof.1.fnc,
-                proof.1.old_value,
-                proof.1.new_value,
-                proof.2.old_root,
-                proof.2.new_root,
-            )
-            .unwrap();
-            assert_eq!(proof.2.fnc, ProcessMerkleProofRole::ProcessInsert);
-
-            proof_t.0.set_witness(pw, &proof.0);
-            proof_t.1.set_witness(pw, &proof.1);
-            proof_t.2.set_witness(pw, &proof.2);
-
-            prev_interior_deposit_digest = proof.0.new_root;
-        }
-        let interior_deposit_digest = prev_interior_deposit_digest;
-
-        let default_proof = SmtProcessProof::with_root(Default::default());
-        let default_proof0 = SmtProcessProof::with_root(interior_deposit_digest);
-        for proof_t in self
-            .deposit_process_proofs
-            .iter()
-            .skip(deposit_process_proofs.len())
-        {
-            proof_t.0.set_witness(pw, &default_proof0);
-            proof_t.1.set_witness(pw, &default_proof);
-            proof_t.2.set_witness(pw, &default_proof);
+            target.set_witness::<_, H, Vec<bool>>(pw, value, true);
         }
 
-        *interior_deposit_digest
+        let default_leaf_data = ContributedAsset::default();
+        for target in self
+            .deposit_process_proof
+            .iter()
+            .skip(value.deposit_process_proof.len())
+        {
+            target.set_witness::<_, H, Vec<bool>>(
+                pw,
+                &PurgeOutputProcessProof {
+                    // siblings: default_merkle_proof.siblings.clone(),
+                    siblings: target
+                        .siblings
+                        .iter()
+                        .map(|_| HashOut::ZERO)
+                        .collect::<Vec<_>>(),
+                    index: target.index.iter().map(|_| false).collect::<Vec<_>>(),
+                    new_leaf_data: default_leaf_data,
+                },
+                false,
+            );
+        }
+
+        value.calculate()
     }
 }
 
-/// Returns `(block_tx_root, old_world_state_root, new_world_state_root)`
-pub fn calc_deposit_digest<F: RichField + Extendable<D>, H: AlgebraicHasher<F>, const D: usize>(
-    builder: &mut CircuitBuilder<F, D>,
-    deposit_process_proofs: &[(
-        SparseMerkleProcessProofTarget,
-        SparseMerkleProcessProofTarget,
-        SparseMerkleProcessProofTarget,
-    )],
-) -> HashOutTarget {
-    let zero = builder.zero();
-    let mut interior_deposit_digest = HashOutTarget {
-        elements: [zero; 4],
-    };
-    for proof_t in deposit_process_proofs {
-        let ProcessMerkleProofRoleTarget {
-            is_insert_or_no_op, ..
-        } = get_process_merkle_proof_role(builder, proof_t.2.fnc);
-        let constant_true = builder._true();
-        builder.connect(is_insert_or_no_op.target, constant_true.target);
-        verify_layered_smt_target_connection(
-            builder,
-            proof_t.0.fnc,
-            proof_t.0.old_value,
-            proof_t.0.new_value,
-            proof_t.1.old_root,
-            proof_t.1.new_root,
-        );
-        verify_layered_smt_target_connection(
-            builder,
-            proof_t.1.fnc,
-            proof_t.1.old_value,
-            proof_t.1.new_value,
-            proof_t.2.old_root,
-            proof_t.2.new_root,
-        );
+// impl DepositBlockProductionTarget {
+//     pub fn add_virtual_to<F: RichField + Extendable<D>, H: AlgebraicHasher<F>, const D: usize>(
+//         builder: &mut CircuitBuilder<F, D>,
+//         log_n_recipients: usize,
+//         log_n_kinds: usize,
+//     ) -> Self {
+//         let n_level = log_n_recipients + log_n_kinds;
+//         let n_deposits = 1 << n_level;
+//         let deposit_list = (0..n_deposits)
+//             .map(|_| ContributedAssetTarget::add_virtual_to(builder))
+//             .collect::<Vec<_>>();
 
-        builder.connect_hashes(proof_t.0.old_root, interior_deposit_digest);
-        interior_deposit_digest = proof_t.0.new_root;
-    }
+//         let leaves = deposit_list
+//             .iter()
+//             .map(|v| builder.hash_n_to_hash_no_pad::<H>(v.encode()))
+//             .collect::<Vec<_>>();
+//         let interior_deposit_digest =
+//             get_merkle_root_target_from_leaves::<_, H, D>(builder, leaves);
 
-    interior_deposit_digest
-}
+//         Self {
+//             deposit_list,
+//             interior_deposit_digest,
+//             log_n_recipients,
+//             log_n_kinds,
+//         }
+//     }
+
+//     /// Returns `interior_deposit_digest`
+//     pub fn set_witness<F: RichField + Extendable<D>, H: AlgebraicHasher<F>, const D: usize>(
+//         &self,
+//         pw: &mut impl Witness<F>,
+//         value: &DepositBlockProduction<F>,
+//     ) -> anyhow::Result<H::Hash> {
+//         for (target, value) in self.deposit_list.iter().zip(value.deposit_list.iter()) {
+//             target.set_witness(pw, *value);
+//         }
+
+//         for target in self.deposit_list.iter().skip(value.deposit_list.len()) {
+//             target.set_witness(pw, ContributedAsset::default());
+//         }
+
+//         value.calculate::<H>()
+//     }
+// }
 
 #[cfg(test)]
 mod tests {
     use std::time::Instant;
 
     use plonky2::{
-        field::types::{Field, Field64},
-        hash::hash_types::HashOut,
         iop::witness::PartialWitness,
         plonk::{
             circuit_builder::CircuitBuilder,
@@ -184,13 +230,12 @@ mod tests {
     };
 
     use crate::{
-        rollup::gadgets::deposit_block::DepositBlockProductionTarget,
-        sparse_merkle_tree::goldilocks_poseidon::{
-            LayeredLayeredPoseidonSparseMerkleTree, NodeDataMemory, RootDataTmp,
+        config::RollupConstants,
+        rollup::{
+            block::make_sample_circuit_inputs,
+            gadgets::deposit_block::{DepositBlockProduction, DepositBlockProductionTarget},
         },
-        transaction::gadgets::deposit_info::DepositInfo,
-        utils::hash::WrappedHashOut,
-        zkdsa::account::{private_key_to_account, Address},
+        transaction::{gadgets::purge::PurgeOutputProcessProof, tree::tx_diff::TxDiffTree},
     };
 
     #[test]
@@ -198,68 +243,91 @@ mod tests {
         const D: usize = 2;
         type C = PoseidonGoldilocksConfig;
         type F = <C as GenericConfig<D>>::F;
-        const LOG_N_RECIPIENTS: usize = 3;
-        const LOG_N_CONTRACTS: usize = 3;
-        const LOG_N_VARIABLES: usize = 3;
-        const N_DEPOSITS: usize = 2;
+        type H = <C as GenericConfig<D>>::InnerHasher;
 
-        let sender2_private_key = HashOut {
-            elements: [
-                F::from_canonical_u64(15657143458229430356),
-                F::from_canonical_u64(6012455030006979790),
-                F::from_canonical_u64(4280058849535143691),
-                F::from_canonical_u64(5153662694263190591),
-            ],
+        const ROLLUP_CONSTANTS: RollupConstants = RollupConstants {
+            log_max_n_users: 3,
+            log_max_n_txs: 3,
+            log_max_n_contracts: 3,
+            log_max_n_variables: 3,
+            log_n_txs: 2,
+            log_n_recipients: 3,
+            log_n_contracts: 3,
+            log_n_variables: 3,
+            n_registrations: 2,
+            n_diffs: 2,
+            n_merges: 2,
+            n_deposits: 2,
+            n_blocks: 2,
         };
-        let sender2_account = private_key_to_account(sender2_private_key);
-        let sender2_address = sender2_account.address.0;
+        let examples = make_sample_circuit_inputs::<C, D>(ROLLUP_CONSTANTS);
 
         let config = CircuitConfig::standard_recursion_config();
         let mut builder = CircuitBuilder::<F, D>::new(config);
-        // builder.debug_gate_row = Some(529); // xors in SparseMerkleProcessProof in DepositBlock
 
         // deposit block
-        let deposit_block_target =
-            DepositBlockProductionTarget::add_virtual_to::<F, <C as GenericConfig<D>>::Hasher, D>(
-                &mut builder,
-                LOG_N_RECIPIENTS,
-                LOG_N_CONTRACTS,
-                LOG_N_VARIABLES,
-                N_DEPOSITS,
-            );
+        let deposit_block_target = DepositBlockProductionTarget::add_virtual_to::<F, H, D>(
+            &mut builder,
+            ROLLUP_CONSTANTS.log_n_recipients,
+            ROLLUP_CONSTANTS.log_n_contracts + ROLLUP_CONSTANTS.log_n_variables,
+        );
         builder.register_public_inputs(&deposit_block_target.interior_deposit_digest.elements);
         let circuit_data = builder.build::<C>();
 
-        let deposit_list = vec![DepositInfo {
-            receiver_address: Address(sender2_address),
-            contract_address: Address(*WrappedHashOut::from_u128(1)),
-            variable_index: 0u8.into(),
-            amount: F::from_noncanonical_u64(1),
-        }];
+        assert_eq!(circuit_data.common.degree_bits(), 15);
 
-        let mut deposit_tree = LayeredLayeredPoseidonSparseMerkleTree::new(
-            NodeDataMemory::default(),
-            RootDataTmp::default(),
+        // let sender2_private_key = HashOut {
+        //     elements: [
+        //         F::from_canonical_u64(15657143458229430356),
+        //         F::from_canonical_u64(6012455030006979790),
+        //         F::from_canonical_u64(4280058849535143691),
+        //         F::from_canonical_u64(5153662694263190591),
+        //     ],
+        // };
+        // let sender2_account = private_key_to_account(sender2_private_key);
+        // let sender2_address = sender2_account.address.0;
+
+        // let deposit_list = vec![ContributedAsset {
+        //     receiver_address: Address(sender2_address),
+        //     kind: TokenKind {
+        //         contract_address: Address(*WrappedHashOut::from_u128(1)),
+        //         variable_index: 0u8.into(),
+        //     },
+        //     amount: 1,
+        // }];
+
+        let deposit_list = &examples[0].deposit_list;
+
+        let mut tx_diff_tree = TxDiffTree::<F, H>::new(
+            ROLLUP_CONSTANTS.log_n_recipients,
+            ROLLUP_CONSTANTS.log_n_contracts + ROLLUP_CONSTANTS.log_n_variables,
         );
-        let deposit_process_proofs = deposit_list
-            .iter()
-            .map(|leaf| {
-                deposit_tree
-                    .set(
-                        leaf.receiver_address.0.into(),
-                        leaf.contract_address.0.into(),
-                        leaf.variable_index.to_hash_out().into(),
-                        HashOut::from_partial(&[leaf.amount]).into(),
-                    )
-                    .unwrap()
-            })
-            .collect::<Vec<_>>();
+
+        let mut deposit_process_proof = vec![];
+        for asset in deposit_list {
+            tx_diff_tree.insert(*asset).unwrap();
+            let proof = tx_diff_tree
+                .prove_leaf_node(&asset.receiver_address, &asset.kind)
+                .unwrap();
+            let process_proof = PurgeOutputProcessProof {
+                siblings: proof.siblings,
+                index: proof.index,
+                new_leaf_data: *asset,
+            };
+            deposit_process_proof.push(process_proof);
+        }
 
         let mut pw = PartialWitness::new();
-        let interior_deposit_digest =
-            deposit_block_target.set_witness::<F, D>(&mut pw, &deposit_process_proofs);
+        let deposit_process_proof = DepositBlockProduction {
+            deposit_process_proof,
+            log_n_recipients: ROLLUP_CONSTANTS.log_n_recipients,
+            log_n_kinds: ROLLUP_CONSTANTS.log_n_contracts + ROLLUP_CONSTANTS.log_n_variables,
+        };
+        let interior_deposit_digest = deposit_block_target
+            .set_witness::<F, H, D>(&mut pw, &deposit_process_proof)
+            .unwrap();
 
-        println!("start proving: block_proof");
+        println!("start proving: deposit_block_proof");
         let start = Instant::now();
         let deposit_block_proof = circuit_data.prove(pw).unwrap();
         let end = start.elapsed();
@@ -273,10 +341,16 @@ mod tests {
         circuit_data.verify(deposit_block_proof).unwrap();
 
         let mut pw = PartialWitness::new();
-        let default_interior_deposit_digest =
-            deposit_block_target.set_witness::<F, D>(&mut pw, &[]);
+        let default_deposit_process_proof = DepositBlockProduction {
+            deposit_process_proof: Default::default(),
+            log_n_recipients: ROLLUP_CONSTANTS.log_n_recipients,
+            log_n_kinds: ROLLUP_CONSTANTS.log_n_contracts + ROLLUP_CONSTANTS.log_n_variables,
+        };
+        let default_interior_deposit_digest = deposit_block_target
+            .set_witness::<F, H, D>(&mut pw, &default_deposit_process_proof)
+            .unwrap();
 
-        println!("start proving: block_proof");
+        println!("start proving: default_deposit_block_proof");
         let start = Instant::now();
         let default_deposit_block_proof = circuit_data.prove(pw).unwrap();
         let end = start.elapsed();
